Renaming lib to tripipy and update imports
lib was conflicting with keywords
imports are fixed to use the library instead of the files
